+++ HISTORY.md
@@ -5,6 +5,8 @@
   resources file. Improves ability to support additional organisms and genome
   builds.
 - Provide paired tumor/normal variant calling with VarScan. Thanks to Luca Beltrame.
+- Require bash shell and use of pipefail for piped commands. Ensures rapid
+  detection of failures during piped steps like alignment.
 - Support for GATK 2.7
 - Pull `description` sample names from BAM files if not present in input
   configuration file. Thanks to Paul Tang for suggestion.